Fix saving images with padding
image sharp can't handle data with line padding in a stride (for
instance given a small dds image or a dds image with an odd width /
height, as dds images are block based). The fix is to create a temporary
buffer with the dimensions that image sharp expects and copy over the
required data.